export methods to a service file
